Rename to octet_stream_json.py
